Still working on conv2_up. Refactored conv2_down.
conv2_down() now accepts the input and output dimensions.

This is important because of the "array[row*width + col]" indexing
scheme used.

I was running into difficulty with conv2_up() because of this very issue
of inflexibility/incompatibliity of the input/output dimensions.

Wrote get_resized_array() and edited buffer_init() so that all
IntermediateTypedArrays() have "width" and "height" attributes.

These are used by conv2_down() (and eventually conv2_up()) to simplify
the calling signature.