fixed conv2_down. 3.5fps
previous commit was partly wrong...

The reason it was fast was because it was running the original code that just copied the edges.
I forgot to remove the old code, and it redefined the function and got run instead.

I'm still surprised the performance dropped so much just from handling edges.