Back at 2 commits ago but with WebAssembly. 14fps.
Basically same functionality as 2 commits ago (7dc511) except implemented with WebAssembly.

Compiled color_space.js to WebAssembly and saw little to no improvement in frame rate.

Compiled blur_filter.js to WebAssembly and saw noticeable improvement (15fps vs. 3fps).

However, the output is a bit wrong. Still need to debug a little.

Just committing because a lot of code compiled from WebAssembly is working.

In worst case might have to learn WebGL as well. 15 fps doesn't look like it will
last through the beating of building a pyramid and reconstructing it...

I had to work on improving the speed because the frame rates were getting too choppy
to be viable for the eventual time filtering step.