Fixed next dimensions logic, sub-pyramids to rgb
The next_pyramid_dimensions did not update if the key existed, which is wrong.
Led to subtle bug due to the formula for "next dimensions" not being unique.
(The ceiling of a division.)

Added color space conversion back to rgb for sub pyramid frames.
Cleaned up some of the display_pyramid() code.

Investigating still why at blur_size=2 there appears to be quadrants
of the resulting image which are distinctly different.
Am I indexing wrong still somehow in the corr_down/corr_up() functions?...
These corr functions are so hard.