Fixed Strange Quadrant Bug. "w <= post".
It was the same bug I found quite a few commits ago concerning including the last point in the kernel in the loop indices.

I had "w < post" as the end condition everywhere (in the corr_up() functions), when it should have been "w <= post".

Also I had as a check for the edge "xx >= operate_width - post" but it should have been "xx >= clip_width - post".
(And analogously for the "yy" version.)

These two bugs were found by logging.

The darkness of the image is because only half of the elements of the kernel are getting used per each convolution pixel, so the brightness is naturally half of what it should be.

I need to consider whether I can just scale the kernel weights by x2 during the corr_up() functions or what...

Also need to fix the YCbCr color space converter. It doesn't seem to be working (either the forward or the inverse, haven't determined, but the result looks green.)

I'm really glad to have found those two indexing bugs that are straightforward to understand.
Need to update the C code to reflect these fixes. Currently only the js versions are corrected.