UI Overhall. Reconstr pyramid. Refactor memory.js
Revamped options pane.

Laid groundwork for time filtering with functions in utils.js and refactor of pyramid memory in memory.js.

Reconstructing the pyramid was very straightforward.

Most of the effort in this commit went towards the UI options pane to makes it more convenient to use than the standard built-in input html elements (see custom_ui_init() in setup.js).

Next up really is just the functionality of time filtering. (IIR bandpass is first.)